@@ -1,10 +1,11 @@
 name: flutter_snowboy
-description: Flutter plugin for Snowboy hotword detection
+description: Flutter wrapper plugin for Snowboy hotword detection. Snowboy is a cross-platform
+DNN-based hotword detection toolkit developed by KITT.AI.
 version: 0.1.1+1
 homepage: https://github.com/mideind/flutter_snowboy
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=2.17.0"
 
 dependencies:
