@@ -1,11 +1,11 @@
 /*
- * Copyright (C) 2021-2022 Miðeind ehf.
+ * Copyright (C) 2021-2023 Miðeind ehf.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ * https://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -18,6 +18,6 @@
 package ai.kitt.snowboy;
 
 public class Constants {
-    public static final String ASSETS_DIRNAME = "snowboy";
-    public static final String COMMON_RES_FILENAME = "common.res";
+  public static final String ASSETS_DIRNAME = "snowboy";
+  public static final String COMMON_RES_FILENAME = "common.res";
 }
