@@ -1,10 +1,11 @@
 /* ----------------------------------------------------------------------------
- * This file was automatically generated by SWIG (http://www.swig.org).
+ * This file was automatically generated by SWIG (https://www.swig.org).
  * Version 4.0.2
  *
  * Do not make changes to this file unless you know what you are doing--modify
  * the SWIG interface file instead.
- * ----------------------------------------------------------------------------- */
+ * -----------------------------------------------------------------------------
+ */
 
 package ai.kitt.snowboy;
 
@@ -45,7 +46,8 @@ public class SnowboyDetect {
   }
 
   public int RunDetection(String data, boolean is_end) {
-    return snowboyJNI.SnowboyDetect_RunDetection__SWIG_0(swigCPtr, this, data, is_end);
+    return snowboyJNI.SnowboyDetect_RunDetection__SWIG_0(swigCPtr, this, data,
+                                                         is_end);
   }
 
   public int RunDetection(String data) {
@@ -53,27 +55,33 @@ public class SnowboyDetect {
   }
 
   public int RunDetection(float[] data, int array_length, boolean is_end) {
-    return snowboyJNI.SnowboyDetect_RunDetection__SWIG_2(swigCPtr, this, data, array_length, is_end);
+    return snowboyJNI.SnowboyDetect_RunDetection__SWIG_2(swigCPtr, this, data,
+                                                         array_length, is_end);
   }
 
   public int RunDetection(float[] data, int array_length) {
-    return snowboyJNI.SnowboyDetect_RunDetection__SWIG_3(swigCPtr, this, data, array_length);
+    return snowboyJNI.SnowboyDetect_RunDetection__SWIG_3(swigCPtr, this, data,
+                                                         array_length);
   }
 
   public int RunDetection(short[] data, int array_length, boolean is_end) {
-    return snowboyJNI.SnowboyDetect_RunDetection__SWIG_4(swigCPtr, this, data, array_length, is_end);
+    return snowboyJNI.SnowboyDetect_RunDetection__SWIG_4(swigCPtr, this, data,
+                                                         array_length, is_end);
   }
 
   public int RunDetection(short[] data, int array_length) {
-    return snowboyJNI.SnowboyDetect_RunDetection__SWIG_5(swigCPtr, this, data, array_length);
+    return snowboyJNI.SnowboyDetect_RunDetection__SWIG_5(swigCPtr, this, data,
+                                                         array_length);
   }
 
   public int RunDetection(int[] data, int array_length, boolean is_end) {
-    return snowboyJNI.SnowboyDetect_RunDetection__SWIG_6(swigCPtr, this, data, array_length, is_end);
+    return snowboyJNI.SnowboyDetect_RunDetection__SWIG_6(swigCPtr, this, data,
+                                                         array_length, is_end);
   }
 
   public int RunDetection(int[] data, int array_length) {
-    return snowboyJNI.SnowboyDetect_RunDetection__SWIG_7(swigCPtr, this, data, array_length);
+    return snowboyJNI.SnowboyDetect_RunDetection__SWIG_7(swigCPtr, this, data,
+                                                         array_length);
   }
 
   public void SetSensitivity(String sensitivity_str) {
@@ -81,7 +89,8 @@ public class SnowboyDetect {
   }
 
   public void SetHighSensitivity(String high_sensitivity_str) {
-    snowboyJNI.SnowboyDetect_SetHighSensitivity(swigCPtr, this, high_sensitivity_str);
+    snowboyJNI.SnowboyDetect_SetHighSensitivity(swigCPtr, this,
+                                                high_sensitivity_str);
   }
 
   public String GetSensitivity() {
@@ -115,5 +124,4 @@ public class SnowboyDetect {
   public int BitsPerSample() {
     return snowboyJNI.SnowboyDetect_BitsPerSample(swigCPtr, this);
   }
-
 }
