@@ -1,10 +1,11 @@
 /* ----------------------------------------------------------------------------
- * This file was automatically generated by SWIG (http://www.swig.org).
+ * This file was automatically generated by SWIG (https://www.swig.org).
  * Version 4.0.2
  *
  * Do not make changes to this file unless you know what you are doing--modify
  * the SWIG interface file instead.
- * ----------------------------------------------------------------------------- */
+ * -----------------------------------------------------------------------------
+ */
 
 package ai.kitt.snowboy;
 
@@ -40,9 +41,7 @@ public class SnowboyVad {
     this(snowboyJNI.new_SnowboyVad(resource_filename), true);
   }
 
-  public boolean Reset() {
-    return snowboyJNI.SnowboyVad_Reset(swigCPtr, this);
-  }
+  public boolean Reset() { return snowboyJNI.SnowboyVad_Reset(swigCPtr, this); }
 
   public int RunVad(String data, boolean is_end) {
     return snowboyJNI.SnowboyVad_RunVad__SWIG_0(swigCPtr, this, data, is_end);
@@ -53,27 +52,33 @@ public class SnowboyVad {
   }
 
   public int RunVad(float[] data, int array_length, boolean is_end) {
-    return snowboyJNI.SnowboyVad_RunVad__SWIG_2(swigCPtr, this, data, array_length, is_end);
+    return snowboyJNI.SnowboyVad_RunVad__SWIG_2(swigCPtr, this, data,
+                                                array_length, is_end);
   }
 
   public int RunVad(float[] data, int array_length) {
-    return snowboyJNI.SnowboyVad_RunVad__SWIG_3(swigCPtr, this, data, array_length);
+    return snowboyJNI.SnowboyVad_RunVad__SWIG_3(swigCPtr, this, data,
+                                                array_length);
   }
 
   public int RunVad(short[] data, int array_length, boolean is_end) {
-    return snowboyJNI.SnowboyVad_RunVad__SWIG_4(swigCPtr, this, data, array_length, is_end);
+    return snowboyJNI.SnowboyVad_RunVad__SWIG_4(swigCPtr, this, data,
+                                                array_length, is_end);
   }
 
   public int RunVad(short[] data, int array_length) {
-    return snowboyJNI.SnowboyVad_RunVad__SWIG_5(swigCPtr, this, data, array_length);
+    return snowboyJNI.SnowboyVad_RunVad__SWIG_5(swigCPtr, this, data,
+                                                array_length);
   }
 
   public int RunVad(int[] data, int array_length, boolean is_end) {
-    return snowboyJNI.SnowboyVad_RunVad__SWIG_6(swigCPtr, this, data, array_length, is_end);
+    return snowboyJNI.SnowboyVad_RunVad__SWIG_6(swigCPtr, this, data,
+                                                array_length, is_end);
   }
 
   public int RunVad(int[] data, int array_length) {
-    return snowboyJNI.SnowboyVad_RunVad__SWIG_7(swigCPtr, this, data, array_length);
+    return snowboyJNI.SnowboyVad_RunVad__SWIG_7(swigCPtr, this, data,
+                                                array_length);
   }
 
   public void SetAudioGain(float audio_gain) {
@@ -95,5 +100,4 @@ public class SnowboyVad {
   public int BitsPerSample() {
     return snowboyJNI.SnowboyVad_BitsPerSample(swigCPtr, this);
   }
-
 }
