@@ -1,12 +1,12 @@
 /* ----------------------------------------------------------------------------
- * This file was automatically generated by SWIG (http://www.swig.org).
+ * This file was automatically generated by SWIG (https://www.swig.org).
  * Version 4.0.2
  *
  * Do not make changes to this file unless you know what you are doing--modify
  * the SWIG interface file instead.
- * ----------------------------------------------------------------------------- */
+ * -----------------------------------------------------------------------------
+ */
 
 package ai.kitt.snowboy;
 
-public class snowboy {
-}
+public class snowboy {}
