@@ -1,49 +1,103 @@
 /* ----------------------------------------------------------------------------
- * This file was automatically generated by SWIG (http://www.swig.org).
+ * This file was automatically generated by SWIG (https://www.swig.org).
  * Version 4.0.2
  *
  * Do not make changes to this file unless you know what you are doing--modify
  * the SWIG interface file instead.
- * ----------------------------------------------------------------------------- */
+ * -----------------------------------------------------------------------------
+ */
 
 package ai.kitt.snowboy;
 
 public class snowboyJNI {
   public final static native long new_SnowboyDetect(String jarg1, String jarg2);
-  public final static native boolean SnowboyDetect_Reset(long jarg1, SnowboyDetect jarg1_);
-  public final static native int SnowboyDetect_RunDetection__SWIG_0(long jarg1, SnowboyDetect jarg1_, String jarg2, boolean jarg3);
-  public final static native int SnowboyDetect_RunDetection__SWIG_1(long jarg1, SnowboyDetect jarg1_, String jarg2);
-  public final static native int SnowboyDetect_RunDetection__SWIG_2(long jarg1, SnowboyDetect jarg1_, float[] jarg2, int jarg3, boolean jarg4);
-  public final static native int SnowboyDetect_RunDetection__SWIG_3(long jarg1, SnowboyDetect jarg1_, float[] jarg2, int jarg3);
-  public final static native int SnowboyDetect_RunDetection__SWIG_4(long jarg1, SnowboyDetect jarg1_, short[] jarg2, int jarg3, boolean jarg4);
-  public final static native int SnowboyDetect_RunDetection__SWIG_5(long jarg1, SnowboyDetect jarg1_, short[] jarg2, int jarg3);
-  public final static native int SnowboyDetect_RunDetection__SWIG_6(long jarg1, SnowboyDetect jarg1_, int[] jarg2, int jarg3, boolean jarg4);
-  public final static native int SnowboyDetect_RunDetection__SWIG_7(long jarg1, SnowboyDetect jarg1_, int[] jarg2, int jarg3);
-  public final static native void SnowboyDetect_SetSensitivity(long jarg1, SnowboyDetect jarg1_, String jarg2);
-  public final static native void SnowboyDetect_SetHighSensitivity(long jarg1, SnowboyDetect jarg1_, String jarg2);
-  public final static native String SnowboyDetect_GetSensitivity(long jarg1, SnowboyDetect jarg1_);
-  public final static native void SnowboyDetect_SetAudioGain(long jarg1, SnowboyDetect jarg1_, float jarg2);
-  public final static native void SnowboyDetect_UpdateModel(long jarg1, SnowboyDetect jarg1_);
-  public final static native int SnowboyDetect_NumHotwords(long jarg1, SnowboyDetect jarg1_);
-  public final static native void SnowboyDetect_ApplyFrontend(long jarg1, SnowboyDetect jarg1_, boolean jarg2);
-  public final static native int SnowboyDetect_SampleRate(long jarg1, SnowboyDetect jarg1_);
-  public final static native int SnowboyDetect_NumChannels(long jarg1, SnowboyDetect jarg1_);
-  public final static native int SnowboyDetect_BitsPerSample(long jarg1, SnowboyDetect jarg1_);
+  public final static native boolean SnowboyDetect_Reset(long jarg1,
+                                                         SnowboyDetect jarg1_);
+  public final static native int
+  SnowboyDetect_RunDetection__SWIG_0(long jarg1, SnowboyDetect jarg1_,
+                                     String jarg2, boolean jarg3);
+  public final static native int
+  SnowboyDetect_RunDetection__SWIG_1(long jarg1, SnowboyDetect jarg1_,
+                                     String jarg2);
+  public final static native int
+  SnowboyDetect_RunDetection__SWIG_2(long jarg1, SnowboyDetect jarg1_,
+                                     float[] jarg2, int jarg3, boolean jarg4);
+  public final static native int
+  SnowboyDetect_RunDetection__SWIG_3(long jarg1, SnowboyDetect jarg1_,
+                                     float[] jarg2, int jarg3);
+  public final static native int
+  SnowboyDetect_RunDetection__SWIG_4(long jarg1, SnowboyDetect jarg1_,
+                                     short[] jarg2, int jarg3, boolean jarg4);
+  public final static native int
+  SnowboyDetect_RunDetection__SWIG_5(long jarg1, SnowboyDetect jarg1_,
+                                     short[] jarg2, int jarg3);
+  public final static native int
+  SnowboyDetect_RunDetection__SWIG_6(long jarg1, SnowboyDetect jarg1_,
+                                     int[] jarg2, int jarg3, boolean jarg4);
+  public final static native int
+  SnowboyDetect_RunDetection__SWIG_7(long jarg1, SnowboyDetect jarg1_,
+                                     int[] jarg2, int jarg3);
+  public final static native void
+  SnowboyDetect_SetSensitivity(long jarg1, SnowboyDetect jarg1_, String jarg2);
+  public final static native void
+  SnowboyDetect_SetHighSensitivity(long jarg1, SnowboyDetect jarg1_,
+                                   String jarg2);
+  public final static native String
+  SnowboyDetect_GetSensitivity(long jarg1, SnowboyDetect jarg1_);
+  public final static native void
+  SnowboyDetect_SetAudioGain(long jarg1, SnowboyDetect jarg1_, float jarg2);
+  public final static native void
+  SnowboyDetect_UpdateModel(long jarg1, SnowboyDetect jarg1_);
+  public final static native int
+  SnowboyDetect_NumHotwords(long jarg1, SnowboyDetect jarg1_);
+  public final static native void
+  SnowboyDetect_ApplyFrontend(long jarg1, SnowboyDetect jarg1_, boolean jarg2);
+  public final static native int SnowboyDetect_SampleRate(long jarg1,
+                                                          SnowboyDetect jarg1_);
+  public final static native int
+  SnowboyDetect_NumChannels(long jarg1, SnowboyDetect jarg1_);
+  public final static native int
+  SnowboyDetect_BitsPerSample(long jarg1, SnowboyDetect jarg1_);
   public final static native void delete_SnowboyDetect(long jarg1);
   public final static native long new_SnowboyVad(String jarg1);
-  public final static native boolean SnowboyVad_Reset(long jarg1, SnowboyVad jarg1_);
-  public final static native int SnowboyVad_RunVad__SWIG_0(long jarg1, SnowboyVad jarg1_, String jarg2, boolean jarg3);
-  public final static native int SnowboyVad_RunVad__SWIG_1(long jarg1, SnowboyVad jarg1_, String jarg2);
-  public final static native int SnowboyVad_RunVad__SWIG_2(long jarg1, SnowboyVad jarg1_, float[] jarg2, int jarg3, boolean jarg4);
-  public final static native int SnowboyVad_RunVad__SWIG_3(long jarg1, SnowboyVad jarg1_, float[] jarg2, int jarg3);
-  public final static native int SnowboyVad_RunVad__SWIG_4(long jarg1, SnowboyVad jarg1_, short[] jarg2, int jarg3, boolean jarg4);
-  public final static native int SnowboyVad_RunVad__SWIG_5(long jarg1, SnowboyVad jarg1_, short[] jarg2, int jarg3);
-  public final static native int SnowboyVad_RunVad__SWIG_6(long jarg1, SnowboyVad jarg1_, int[] jarg2, int jarg3, boolean jarg4);
-  public final static native int SnowboyVad_RunVad__SWIG_7(long jarg1, SnowboyVad jarg1_, int[] jarg2, int jarg3);
-  public final static native void SnowboyVad_SetAudioGain(long jarg1, SnowboyVad jarg1_, float jarg2);
-  public final static native void SnowboyVad_ApplyFrontend(long jarg1, SnowboyVad jarg1_, boolean jarg2);
-  public final static native int SnowboyVad_SampleRate(long jarg1, SnowboyVad jarg1_);
-  public final static native int SnowboyVad_NumChannels(long jarg1, SnowboyVad jarg1_);
-  public final static native int SnowboyVad_BitsPerSample(long jarg1, SnowboyVad jarg1_);
+  public final static native boolean SnowboyVad_Reset(long jarg1,
+                                                      SnowboyVad jarg1_);
+  public final static native int SnowboyVad_RunVad__SWIG_0(long jarg1,
+                                                           SnowboyVad jarg1_,
+                                                           String jarg2,
+                                                           boolean jarg3);
+  public final static native int
+  SnowboyVad_RunVad__SWIG_1(long jarg1, SnowboyVad jarg1_, String jarg2);
+  public final static native int
+  SnowboyVad_RunVad__SWIG_2(long jarg1, SnowboyVad jarg1_, float[] jarg2,
+                            int jarg3, boolean jarg4);
+  public final static native int SnowboyVad_RunVad__SWIG_3(long jarg1,
+                                                           SnowboyVad jarg1_,
+                                                           float[] jarg2,
+                                                           int jarg3);
+  public final static native int
+  SnowboyVad_RunVad__SWIG_4(long jarg1, SnowboyVad jarg1_, short[] jarg2,
+                            int jarg3, boolean jarg4);
+  public final static native int SnowboyVad_RunVad__SWIG_5(long jarg1,
+                                                           SnowboyVad jarg1_,
+                                                           short[] jarg2,
+                                                           int jarg3);
+  public final static native int
+  SnowboyVad_RunVad__SWIG_6(long jarg1, SnowboyVad jarg1_, int[] jarg2,
+                            int jarg3, boolean jarg4);
+  public final static native int SnowboyVad_RunVad__SWIG_7(long jarg1,
+                                                           SnowboyVad jarg1_,
+                                                           int[] jarg2,
+                                                           int jarg3);
+  public final static native void
+  SnowboyVad_SetAudioGain(long jarg1, SnowboyVad jarg1_, float jarg2);
+  public final static native void
+  SnowboyVad_ApplyFrontend(long jarg1, SnowboyVad jarg1_, boolean jarg2);
+  public final static native int SnowboyVad_SampleRate(long jarg1,
+                                                       SnowboyVad jarg1_);
+  public final static native int SnowboyVad_NumChannels(long jarg1,
+                                                        SnowboyVad jarg1_);
+  public final static native int SnowboyVad_BitsPerSample(long jarg1,
+                                                          SnowboyVad jarg1_);
   public final static native void delete_SnowboyVad(long jarg1);
 }
